+++ TODO.md
@@ -1,0 +1,6 @@
+# Illuminatrix
+
+## TODO list
+- [ ] Get rid of SystemIdentifier and just TowerEnum directly
+- [ ] Make the initialization and lifecycle of systems consistent
+
