+++ systems/concrete/PrintSoundSystem.py
@@ -1,5 +1,5 @@
-from constants import SystemIdentifier
-from SoundSystem import SoundSystem
+from bases.SoundSystem import SoundSystem
+from constants.constants import SystemIdentifier
 
 
 class PrintSoundSystem(SoundSystem):
