+++ utils/KBHit.py
@@ -20,7 +20,6 @@ GNU General Public License for more details.
 '''
 
 import os
-import time
 from typing import Self
 
 # Windows
@@ -29,9 +28,9 @@ if os.name == 'nt':
 
 # Posix (Linux, OS X)
 else:
+    import atexit
     import sys
     import termios
-    import atexit
     from select import select
